+++ src/fabric/NativePlaidLinkModuleAndroid.ts
@@ -1,7 +1,7 @@
 // we use Object type because methods on the native side use NSDictionary and ReadableMap
 // and we want to stay compatible with those
 import {TurboModuleRegistry, TurboModule} from 'react-native';
-import {Int32} from 'react-native/Libraries/Types/CodegenTypes';
+import {Double} from 'react-native/Libraries/Types/CodegenTypes';
 import {UnsafeObject} from './fabricUtils';
 import {LinkSuccess, LinkExit} from '../Types';
 
@@ -20,7 +20,7 @@ export interface Spec extends TurboModule {
     ): void;
     // those two are here for event emitter methods
     addListener(eventName: string): void;
-    removeListeners(count: Int32): void;
+    removeListeners(count: Double): void;
 }
 
 export default TurboModuleRegistry.get<Spec>('PlaidAndroid');
